Allow more controlled selection of default image.
The default image for executorless (daemon/command) and executorful tasks is
different, now -- the OS distro:version pair is still used but executorful
tasks prepend the "libmesos" account, where we'll post images that have
Mesos libraries pre-installed.

A "docker.index" section has been added to the config file so that users can
specify an index host and the names of accounts to use when following
defaulting rules for executorful and executorless images.